Create IVirtualFolderManager
