+++ MediaBrowser.Controller/Library/ILibraryManager.cs
@@ -16,7 +16,6 @@ using MediaBrowser.Controller.Resolvers;
 using MediaBrowser.Controller.Sorting;
 using MediaBrowser.Model.Configuration;
 using MediaBrowser.Model.Dto;
-using MediaBrowser.Model.Entities;
 using MediaBrowser.Model.IO;
 using MediaBrowser.Model.Querying;
 using Episode = MediaBrowser.Controller.Entities.TV.Episode;
@@ -152,14 +151,6 @@ namespace MediaBrowser.Controller.Library
         Task ValidateMediaLibrary(IProgress<double> progress, CancellationToken cancellationToken);
 
         Task UpdateImagesAsync(BaseItem item, bool forceUpdate = false);
-
-        /// <summary>
-        /// Gets the default view.
-        /// </summary>
-        /// <returns>IEnumerable{VirtualFolderInfo}.</returns>
-        List<VirtualFolderInfo> GetVirtualFolders();
-
-        List<VirtualFolderInfo> GetVirtualFolders(bool includeRefreshState);
 
         /// <summary>
         /// Gets the item by id.
@@ -549,16 +540,6 @@ namespace MediaBrowser.Controller.Library
 
         Guid GetMusicGenreId(string name);
 
-        Task AddVirtualFolder(string name, CollectionTypeOptions? collectionType, LibraryOptions options, bool refreshLibrary);
-
-        Task RemoveVirtualFolder(string name, bool refreshLibrary);
-
-        void AddMediaPath(string virtualFolderName, MediaPathInfo mediaPath);
-
-        void UpdateMediaPath(string virtualFolderName, MediaPathInfo mediaPath);
-
-        void RemoveMediaPath(string virtualFolderName, string mediaPath);
-
         QueryResult<(BaseItem Item, ItemCounts ItemCounts)> GetGenres(InternalItemsQuery query);
 
         QueryResult<(BaseItem Item, ItemCounts ItemCounts)> GetMusicGenres(InternalItemsQuery query);
@@ -576,6 +557,8 @@ namespace MediaBrowser.Controller.Library
         Task RunMetadataSavers(BaseItem item, ItemUpdateType updateReason);
 
         BaseItem GetParentItem(Guid? parentId, Guid? userId);
+
+        Task ValidateTopLibraryFolders(CancellationToken cancellationToken);
 
         /// <summary>
         /// Queue a library scan.
